refactor: 물품보관소 장소 리스트 응답 값 변경
